fixed 10.3 full install issue

+++ scripts/os_settings.sh
@@ -10,9 +10,11 @@ defaults -currentHost write com.apple.screensaver idleTime 0
 pkill Feedback || true
 
 # cleanup ssh enablement from install process
-sudo launchctl unload -w /Library/LaunchDaemons/ssh.plist
-sudo rm /Library/LaunchDaemons/ssh.plist
-#sudo /usr/sbin/systemsetup -f -setremotelogin on 
-sudo launchctl load -w /System/Library/LaunchDaemons/ssh.plist
+if [[ -e /Library/LaunchDaemons/ssh.plist ]]; then
+  sudo launchctl unload -w /Library/LaunchDaemons/ssh.plist
+  sudo rm /Library/LaunchDaemons/ssh.plist
+  #sudo /usr/sbin/systemsetup -f -setremotelogin on
+  sudo launchctl load -w /System/Library/LaunchDaemons/ssh.plist
+fi
 
 exit 0
